Master

+++ src/main.c
@@ -9,24 +9,24 @@ int main(void){
     Stack stack = new_Stack(int);
     Queue queue = new_Queue(int);
     Deque deque = new_Deque(int);
-    short* value = (short*)malloc(sizeof(int) * 10);
+    int* value = (int*)malloc(sizeof(int) * 10);
 
     srand(time(NULL));
 
     for (int i = 0; i < 10; i++) {
         *(value + i) = rand() % 100;
         stack.push(&stack, value + i);
-        for (int j = 0; j < stack.top; j++) System.out.printf("%d ", *(short*)*(stack.data + j));
+        for (int j = 0; j < stack.top; j++) System.out.printf("%d ", *(int*)*(stack.data + j));
         System.out.println("");
     }
 
     for (int i = 0; i < 10; i++) {
-        short* poppedValue = (short*)stack.pop(&stack);
+        int* poppedValue = (int*)stack.pop(&stack);
         if (poppedValue) {
             System.out.println("Popped: %d", *poppedValue);
             free(poppedValue);
         }
-        for (int j = 0; j < stack.top; j++) System.out.printf("%d ", *(short*)*(stack.data + j));
+        for (int j = 0; j < stack.top; j++) System.out.printf("%d ", *(int*)*(stack.data + j));
         System.out.println("");
     }
 
@@ -35,17 +35,17 @@ int main(void){
     for (int i = 0; i < 10; i++) {
         *(value + i) = rand() % 100;
         queue.push(&queue, value + i);
-        for (int j = 0; j < queue.size; j++) System.out.printf("%d ", *(short*)*(queue.data + j));
+        for (int j = 0; j < queue.size; j++) System.out.printf("%d ", *(int*)*(queue.data + j));
         System.out.println("");
     }
 
     for (int i = 0; i < 10; i++) {
-        short* poppedValue = (short*)queue.pop(&queue);
+        int* poppedValue = (int*)queue.pop(&queue);
         if (poppedValue) {
             System.out.println("Popped: %d", *poppedValue);
             free(poppedValue);
         }
-        for (int j = 0; j < queue.size; j++) System.out.printf("%d ", *(short*)*(queue.data + j));
+        for (int j = 0; j < queue.size; j++) System.out.printf("%d ", *(int*)*(queue.data + j));
         System.out.println("");
     }
 
@@ -54,17 +54,17 @@ int main(void){
     for (int i = 0; i < 10; i++) {
         *(value + i) = rand() % 100;
         deque.pushFront(&deque, value + i);
-        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(short*)*(deque.data + j));
+        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(int*)*(deque.data + j));
         System.out.println("");
     }
 
     for (int i = 0; i < 10; i++) {
-        short* poppedValue = (short*)deque.popFront(&deque);
+        int* poppedValue = (int*)deque.popFront(&deque);
         if (poppedValue) {
             System.out.println("Popped: %d", *poppedValue);
             free(poppedValue);
         }
-        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(short*)*(deque.data + j));
+        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(int*)*(deque.data + j));
         System.out.println("");
     }
 
@@ -73,17 +73,17 @@ int main(void){
     for (int i = 0; i < 10; i++) {
         *(value + i) = rand() % 100;
         deque.pushBack(&deque, value + i);
-        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(short*)*(deque.data + j));
+        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(int*)*(deque.data + j));
         System.out.println("");
     }
 
     for (int i = 0; i < 10; i++) {
-        short* poppedValue = (short*)deque.popBack(&deque);
+        int* poppedValue = (int*)deque.popBack(&deque);
         if (poppedValue) {
             System.out.println("Popped: %d", *poppedValue);
             free(poppedValue);
         }
-        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(short*)*(deque.data + j));
+        for (int j = 0; j < deque.size; j++) System.out.printf("%d ", *(int*)*(deque.data + j));
         System.out.println("");
     }
